backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==46.0.2
defusedxml==0.7.1
Django==3.2
django-cors-headers==3.13.0
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
flake8==7.3.0
fpdf==1.7.2
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
oauthlib==3.3.1
pillow==11.3.0
psycopg2-binary==2.9.3
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2025.2
reportlab==3.6.12
requests==2.32.5
requests-oauthlib==2.0.0
shortuuid==1.0.13
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.8.1
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
